Raw and Normalized AUC
AUC score used as it encodes distribution shape and location on log (can be changed to linear) channel value. Input is fca_readfcs-processed FCS 2.0 files organized in a struct (see below for format).

[Column 1]  [Column 2]  [Column 3]   [Column 4]  [Column 5]
    Time         Well label    Parameter      Data           Event count
                                          name

Data is an mxn matrix wherein m is the number of events in the nth parameter. Identities of n are in column 3 of the struct.